remove changeSettings and show popover in Rotor component
